Refatoração (limpeza do interpretador) (Core)
